# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CommunityPost:
    properties:
      anon_name:
        type: string
      category:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_urls:
        items:
          type: string
        type: array
      is_flagged:
        type: boolean
      replies:
        items:
          $ref: '#/definitions/models.CommunityReply'
        type: array
      reply_count:
        type: integer
      title:
        type: string
      user_email:
        type: string
      user_name:
        description: Populated at query time for admin endpoints only; not stored
          in DB.
        type: string
    type: object
  models.CommunityReply:
    properties:
      anon_name:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_flagged:
        type: boolean
      post_id:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: MarketKit
  description: REST API for the MarketKit marketplace platform. Admin routes require
    a Bearer JWT token. User routes require a separate User Bearer JWT token.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: MarketKit API
  version: "1.0"
paths:
...
  /user/community/posts:
    get:
      parameters:
      - description: Category filter
        in: query
        name: category
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CommunityPost'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: List community posts
      tags:
      - User Community
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Post title
        in: formData
        name: title
        required: true
        type: string
      - description: Post content
        in: formData
        name: content
        required: true
        type: string
      - description: 'Category (default: GENERAL)'
        in: formData
        name: category
        type: string
      - description: Up to 3 images
        in: formData
        name: images
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CommunityPost'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Create a community post
      tags:
      - User Community
  /user/community/posts/{id}:
    get:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CommunityPost'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Get a community post with replies
      tags:
      - User Community
  /user/community/posts/{id}/replies:
    get:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CommunityReply'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: List replies for a community post
      tags:
      - User Community
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Reply content
        in: formData
        name: content
        required: true
        type: string
      - description: Optional image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CommunityReply'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserAuth: []
      summary: Reply to a community post
      tags:
      - User Community
...
securityDefinitions:
  AdminAuth:
    description: Admin JWT — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    description: User JWT — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"